requirements.txt
pandas 
mlflow
notebook
numpy
comet-ml==3.49.8
tensorflow==2.19.0
torch
torchvision
scikit-learn
scikit-image
opencv-python
matplotlib
python-box==6.0.2
pyYAML
tqdm
ensure
joblib
types-PyYAML
Flask
Flask-Cors
-e .